client/requirements.txt
# Mac client dependencies for voice input
sounddevice>=0.5.0
websockets>=12.0
pyyaml>=6.0
pyobjc-framework-Cocoa>=10.0
